Update GO_parser_plotter_v1.2.py
Version 1.2 introduces a new command line argument -c, --color, which allows the user to specify a colormap palette for their heatmap given that it is supported by matplotlib and seaborn modules. Version 1.2 also makes the graphing portion of the script more readable and efficient, as well as updated help and usage information.